Fix just updating the NCS
The previous implementation sometimes corrupted the installation. The
new implementation does respect whether a new NCS is installed or
whether it is just updated. In the latter case, a possibly existing
directory `.west` must not be removed and the script `ncsmgr init-ncs`
should not call `west init`. For this, the new implementation requires
an updated version of the toolchain, that honors passing the parameter
`--just-update` to the script `ncsmgr init-ncs`. This update is
implemented in https://github.com/NordicPlayground/toolchain/pull/26.

This new implementation follows the logic as described in
https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/nrf/gs_installing.html#updating-the-repositories

+++ src/Manager/Environment/environmentEffects.js
@@ -135,13 +135,15 @@ const installToolchain = async (dispatch, version, toolchain, toolchainDir) => {
 
 export const isWestPresent = toolchainDir => fs.existsSync(path.resolve(toolchainDir, '../.west/config'));
 
-export const cloneNcs = async (dispatch, version, toolchainDir) => {
+export const cloneNcs = async (dispatch, version, toolchainDir, justUpdate) => {
     dispatch(startCloningSdk(version));
 
-    await fse.remove(path.resolve(path.dirname(toolchainDir), '.west'));
+    if (!justUpdate) {
+        await fse.remove(path.resolve(path.dirname(toolchainDir), '.west'));
+    }
     try {
         const gitBash = path.resolve(toolchainDir, 'git-bash.exe');
-        const initScript = 'unset ZEPHYR_BASE; toolchain/ncsmgr/ncsmgr init-ncs; sleep 3';
+        const initScript = `unset ZEPHYR_BASE; toolchain/ncsmgr/ncsmgr init-ncs ${justUpdate ? '--just-update' : ''}; sleep 3`;
         await promisify(exec)(`"${gitBash}" -c "${initScript}"`);
     } catch (error) {
         console.error(`Failed to clone NCS with error: ${error}`);
@@ -150,7 +152,7 @@ export const cloneNcs = async (dispatch, version, toolchainDir) => {
     dispatch(finishCloningSdk(version, isWestPresent(toolchainDir)));
 };
 
-export const install = async (dispatch, { version, toolchains }) => {
+export const install = async (dispatch, { version, toolchains }, justUpdate) => {
     const toolchain = getLatestToolchain(toolchains);
     const toolchainDir = path.resolve(installDir(), version, 'toolchain');
 
@@ -161,7 +163,7 @@ export const install = async (dispatch, { version, toolchains }) => {
     setHasInstalledAnNcs();
 
     await installToolchain(dispatch, version, toolchain, toolchainDir);
-    await cloneNcs(dispatch, version, toolchainDir);
+    await cloneNcs(dispatch, version, toolchainDir, justUpdate);
 };
 
 export const remove = async (dispatch, { toolchainDir, version }) => {
